refactor: Turn new tracks calculation into hook to prevent repeated updates
