+++ src/components/Tabs/Lineage/types.ts
@@ -1,4 +1,4 @@
-import * as d3 from "d3";
+import type * as d3 from "d3";
 
 export type TrackInfo = {
   id: number;
